Move dataExtend widget to simulation. Add load shapefile function
