+++ library/Mag/ProblemSetter.py
@@ -174,87 +174,6 @@ def setSyntheticProblem(
     return survey
 
 
-def setDataExtentWidget(survey):
-    """
-        Small application to carve out a subset of a larger data set
-    """
-
-    def dataSelector(East, North, Width, Height):
-
-        lims = np.r_[
-            East-Width/2, East+Width/2,
-            North-Height/2, North+Height/2
-        ]
-
-        fig, axs = plt.figure(figsize=(12, 6)), plt.subplot(1, 2, 1)
-        Simulator.plotData2D(
-            xLoc, yLoc, survey.values, marker=False, fig=fig, ax=axs,
-            colorbar=False
-        )
-
-        # axs.scatter(East, North, 20, marker='+')
-        axs.add_patch(
-            Rectangle(
-                (East-Width/2, North-Height/2),
-                Width,
-                Height,
-                facecolor='none', edgecolor='k',
-                zorder=3
-                )
-            )
-
-        # Extract data within window and plot
-        indx = np.logical_and(xLoc > lims[0], xLoc < lims[1])
-
-        indy = np.logical_and(yLoc > lims[2], yLoc < lims[3])
-
-        nx, ny = np.count_nonzero(indx), np.count_nonzero(indy)
-
-        # Create new dataGrid object
-        dataSub = DataIO.dataGrid()
-        dataSub.limits = lims
-        # coordinate_system = grid.coordinate_system
-        dataSub.values = survey.values[:, indx]
-        dataSub.values = dataSub.values[indy, :]
-
-        dataSub.nx, dataSub.ny = nx, ny
-        dataSub.dx, dataSub.dy = survey.dx, survey.dy
-        dataSub.x0, dataSub.y0 = East-Width/2, North-Height/2
-
-        # fig,
-        axs = plt.subplot(1, 2, 2)
-        fig, im, cbar = Simulator.plotData2D(
-            xLoc[indx], yLoc[indy], dataSub.values, marker=False, fig=fig, ax=axs
-        )
-        cbar.set_label('TMI (nT)')
-        return dataSub
-
-    if isinstance(survey, DataIO.dataGrid):
-
-        xLoc = np.asarray(range(survey.nx))*survey.dx+survey.x0
-        yLoc = np.asarray(range(survey.ny))*survey.dy+survey.y0
-        xlim = survey.limits[:2]
-        ylim = survey.limits[2:]
-
-    else:
-        print("Only implemented for class 'dataGrid'")
-        # xLoc = survey.srcField.rxList[0].locs[:, 0]
-        # yLoc = survey.srcField.rxList[0].locs[:, 1]
-        # xlim = np.asarray([xLoc.min(), xLoc.max()])
-        # ylim = np.asarray([yLoc.min(), yLoc.max()])
-        # data = survey.dobs
-
-    out = widgets.interactive(
-            dataSelector,
-            East=widgets.FloatSlider(min=xlim[0], max=xlim[1], step=500, value=669500, continuous_update=False),
-            North=widgets.FloatSlider(min=ylim[0], max=ylim[1], step=10, value=6069500, continuous_update=False),
-            Width=widgets.FloatSlider(min=1000, max=100000, step=1000, value=30000, continuous_update=False),
-            Height=widgets.FloatSlider(min=1000, max=100000, step=1000, value=30000, continuous_update=False)
-            )
-
-    return out
-
-
 def meshBuilder(xyz, h, padDist, meshGlobal=None,
                 expFact=1.3,
                 meshType='TENSOR',
@@ -463,3 +382,8 @@ def refineTree(mesh, xyz, finalize=False, dtype="point", nCpad=[1, 1, 1]):
         NotImplementedError("Only dtype='points' has been implemented")
 
     return mesh
+
+
+# if __name__ == '__main__':
+
+
