+++ library/Mag/Simulator.py
@@ -22,6 +22,7 @@ from library.graphics import graphics
 from matplotlib.ticker import FormatStrFormatter
 import matplotlib as mpl
 from skimage import exposure
+from matplotlib.patches import Rectangle
 import PIL
 
 def PFSimulator(prism, survey):
@@ -1536,26 +1537,10 @@ def gridTilt2Depth(
 
     return out
 
+
 def worldViewerWidget(worldFile, data, grid, z=0):
 
-    world = shapefile.Reader(worldFile)
-    # Extract lines from shape file
-    X, Y = [], []
-    for shape in world.shapeRecords():
-
-        for ii, part in enumerate(shape.shape.parts):
-
-            if ii != len(shape.shape.parts)-1:
-                x = [i[0] for i in shape.shape.points[shape.shape.parts[ii]:shape.shape.parts[ii+1]:50]]
-                y = [i[1] for i in shape.shape.points[shape.shape.parts[ii]:shape.shape.parts[ii+1]:50]]
-
-            else:
-                x = [i[0] for i in shape.shape.points[shape.shape.parts[ii]::50]]
-                y = [i[1] for i in shape.shape.points[shape.shape.parts[ii]::50]]
-
-            if len(x) > 10:
-                X.append(np.vstack(x))
-                Y.append(np.vstack(y))
+    X, Y = DataIO.readShapefile(worldFile)
 
     def plotCountry(X, Y, ax=None, fill=True, linewidth=1):
 
@@ -1595,7 +1580,6 @@ def worldViewerWidget(worldFile, data, grid, z=0):
 
         axs = plt.subplot(1, 2, 2)
         axs = plotCountry(X, Y, ax=axs, fill=False)
-        axs.scatter(dataVals[1], dataVals[0], s=50, c='r', marker='s', )
         # axs.set_axis_off()
         axs.set_aspect('equal')
         pos = axs.get_position()
@@ -1604,6 +1588,7 @@ def worldViewerWidget(worldFile, data, grid, z=0):
         # xydata = np.loadtxt("./assets/country-capitals.csv", delimiter=",")
         for key, entry in zip(list(data.keys()), list(data.values())):
             axs.scatter(entry[1], entry[0], c='k')
+        axs.scatter(dataVals[1], dataVals[0], s=50, c='r', marker='s', )
 
         axs.set_title(
           "Earth's Field: " + str(int(dataVals[-3])) + "nT, "
@@ -1791,4 +1776,85 @@ def dataGriddingWidget(survey, EPSGCode=26909, saveAs='DataGrid'):
                                   icon='check'
                                 )
                               )
+    return out
+
+
+def setDataExtentWidget(survey):
+    """
+        Small application to carve out a subset of a larger data set
+    """
+
+    def dataSelector(East, North, Width, Height):
+
+        lims = np.r_[
+            East-Width/2, East+Width/2,
+            North-Height/2, North+Height/2
+        ]
+
+        fig, axs = plt.figure(figsize=(12, 6)), plt.subplot(1, 2, 1)
+        plotData2D(
+            xLoc, yLoc, survey.values, marker=False, fig=fig, ax=axs,
+            colorbar=False
+        )
+
+        # axs.scatter(East, North, 20, marker='+')
+        axs.add_patch(
+            Rectangle(
+                (East-Width/2, North-Height/2),
+                Width,
+                Height,
+                facecolor='none', edgecolor='k',
+                zorder=3
+                )
+            )
+
+        # Extract data within window and plot
+        indx = np.logical_and(xLoc > lims[0], xLoc < lims[1])
+
+        indy = np.logical_and(yLoc > lims[2], yLoc < lims[3])
+
+        nx, ny = np.count_nonzero(indx), np.count_nonzero(indy)
+
+        # Create new dataGrid object
+        dataSub = DataIO.dataGrid()
+        dataSub.limits = lims
+        # coordinate_system = grid.coordinate_system
+        dataSub.values = survey.values[:, indx]
+        dataSub.values = dataSub.values[indy, :]
+
+        dataSub.nx, dataSub.ny = nx, ny
+        dataSub.dx, dataSub.dy = survey.dx, survey.dy
+        dataSub.x0, dataSub.y0 = East-Width/2, North-Height/2
+
+        # fig,
+        axs = plt.subplot(1, 2, 2)
+        fig, im, cbar = plotData2D(
+            xLoc[indx], yLoc[indy], dataSub.values, marker=False, fig=fig, ax=axs
+        )
+        cbar.set_label('TMI (nT)')
+        return dataSub
+
+    if isinstance(survey, DataIO.dataGrid):
+
+        xLoc = np.asarray(range(survey.nx))*survey.dx+survey.x0
+        yLoc = np.asarray(range(survey.ny))*survey.dy+survey.y0
+        xlim = survey.limits[:2]
+        ylim = survey.limits[2:]
+
+    else:
+        print("Only implemented for class 'dataGrid'")
+        # xLoc = survey.srcField.rxList[0].locs[:, 0]
+        # yLoc = survey.srcField.rxList[0].locs[:, 1]
+        # xlim = np.asarray([xLoc.min(), xLoc.max()])
+        # ylim = np.asarray([yLoc.min(), yLoc.max()])
+        # data = survey.dobs
+
+    out = widgets.interactive(
+            dataSelector,
+            East=widgets.FloatSlider(min=xlim[0], max=xlim[1], step=500, value=669500, continuous_update=False),
+            North=widgets.FloatSlider(min=ylim[0], max=ylim[1], step=10, value=6069500, continuous_update=False),
+            Width=widgets.FloatSlider(min=1000, max=100000, step=1000, value=30000, continuous_update=False),
+            Height=widgets.FloatSlider(min=1000, max=100000, step=1000, value=30000, continuous_update=False)
+            )
+
     return out
